revises dev examples for working with stratified datasets
The last few lines joining_groups demonstrates how to call mk_stat
in dplyr with intended effects. There are two main challenges:
    1. using global environment variables in dplyr (I'm avoiding this)
    2. employing mk_stat in the summarize function
We can exploit some known characteristics of the population sizes or
sampling weights to construct a population stat N in the table summary,
then push it with other function arguments into mk_stat to achieve the
expected effect. This will hide some technicality from the user. An
alternative it to reimplement most of the code from mk_stat in make_summary
as it was written. In the interest of not repeating code, we'll take the
steps described briefly in this commit and the top of the source for
make_summary.